requirements.txt
flask
python-dotenv
google-genai
yfinance
pandas
numpy
matplotlib
networkx
torch
torch-geometric
scikit-learn